Broken boost filename hack
The boost date/time library is forcing a link to a file named
boost_chrono.dll which is excludes the compiler name that is
included with the default boost build. The boost compiler action
is already generating libraries with this naming so forcing the
updated filename isn't needed. Add a file exists check so it can
still happen on other systems but be skipped when not required.