Incorrect permissions on Java library files
The Builder files are unzipped with 0000 permissions on all files except for the WISE_Builder_lib folder which has 0775. The permission on the main jar file (WISE_Builder.jar) was being updated but the permissions on the dependencies was not. This stopped Java from being able to load them at runtime which caused a ClassNotFoundException on the first class that Java came across that is contained in those libraries, which happened to be JAXBException. A recursive permission update on the WISE_Builder_lib directory should get it running for WISE-Developers/Project_Issues#182.